✨ Implementar Plan Mejoras Financieras (PASO 1-8)
Fase completa de mejoras de usabilidad y educación financiera para propietario sin formación contable.

## PASO 1: Endpoint /api/semaforo
- Calcula salud financiera mensual (ingresos vs gastos)
- Retorna color (success/warning/danger) e tooltip con 3 valores
- Disponible solo para roles admin/vendedor

## PASO 2: Rutas /reportes/financiero
- GET /reportes/financiero: Vista Estado Financiero Simplificado
- GET /reportes/financiero/pdf: Exportación PDF para análisis
- Incluye helper function obtenerDatosFinancieros() con 8 queries

## PASO 3: Nueva vista financiero.ejs
- Card grande con resultado neto + conclusión automática
- 4 KPIs: Ingresos, Gastos Negocio, Gastos Personales, Pedidos
- Tabla detallada de gastos por categoría con porcentajes
- Cards de Cuentas por Cobrar y Deuda en Tarjetas

## PASO 4: Rediseño reportes/inicio.ejs
- Panel educativo superior con explicación y prioridades
- 6 tarjetas (de 5): Estado Financiero, Balance, Clientes, CxC, Productos, Entradas/Salidas
- Badges de frecuencia recomendada
- Botón ayuda con offcanvas contextual

## PASO 5: Estados vacíos mejorados (4 vistas finanzas)
- gastos-fijos.ejs: Emoji 📅 + explicación de gastos fijos
- compras-cuotas.ejs: Emoji 🛒 + explicación de deuda comprometida
- facturas-recibidas.ejs: Rediseño COMPLETO (🧾 + panel educativo azul + modal tipo factura + resumen mensual + alerta automática)
- vencimientos-fiscales.ejs: Emoji 📆 + explicación de obligaciones fiscales

## PASO 6: Botones de ayuda (offcanvas) - 13 páginas
Patrón estándar en todas las páginas financieras:
- Botón "¿Cómo uso esta sección?" en header
- Offcanvas con 4 secciones: Para qué sirve, Qué cargar y cuándo, Errores comunes, Ejemplo concreto
- Lenguaje simple, ejemplos específicos de imprenta

Páginas actualizadas:
✓ views/reportes/inicio.ejs
✓ views/gastos/index.ejs
✓ views/finanzas/gastos-fijos.ejs
✓ views/finanzas/compras-cuotas.ejs
✓ views/finanzas/facturas-recibidas.ejs
✓ views/finanzas/vencimientos-fiscales.ejs
✓ views/deudas/tarjetas.ejs
✓ views/deudas/cheques.ejs
✓ views/deudas/prestamos.ejs
✓ views/deudas/proveedores-deuda.ejs

## PASO 7: Popovers en formularios de deudas
- Inicialización global en layout.ejs
- Íconos ℹ️ junto a labels de campos críticos
- Popovers informativos sin jerga contable

Campos cubiertos:
- Cheques: número, banco, beneficiario, vencimiento
- Préstamos: entidad, monto original, cuota, tasa, vencimiento mensual
- Proveedores: concepto, monto, vencimiento

## PASO 8: Semáforo financiero en navbar
- CSS: 3 estados (verde/amarillo/rojo) con glow effects
- HTML: Dot condicionado a admin/vendedor
- JS: Fetch /api/semaforo + actualización dinámica + Bootstrap Tooltip

## Beneficios
- Dueño sin contabilidad entiende su salud financiera
- Guía de frecuencia recomendada para revisiones
- Ejemplos concretos de imprenta en todos los help panels
- Validación automática y cálculos de IVA
- Estados vacíos explican qué falta cargar

+++ controllers/catalogoController.js
@@ -51,11 +51,11 @@ module.exports = (db) => {
             
             await db.run(
                 'INSERT INTO catalogo_productos (codigo, nombre, tipo, precio_base, minimo) VALUES (?, ?, ?, ?, ?)',
-                codigo || null, 
-                nombre.trim(), 
-                tipo, 
-                parseFloat(precio_base) || 0, 
-                parseInt(minimo) || 1
+                codigo || null,
+                nombre.trim(),
+                tipo,
+                parseFloat(precio_base) || 0,
+                parseFloat(minimo) || 1
             );
             req.flash('success', 'Producto creado correctamente');
             res.redirect('/catalogo');
@@ -102,10 +102,10 @@ module.exports = (db) => {
             await db.run(
                 'UPDATE catalogo_productos SET codigo = ?, nombre = ?, tipo = ?, precio_base = ?, minimo = ? WHERE id = ?',
                 codigo || null,
-                nombre.trim(), 
-                tipo, 
-                parseFloat(precio_base) || 0, 
-                parseInt(minimo) || 1, 
+                nombre.trim(),
+                tipo,
+                parseFloat(precio_base) || 0,
+                parseFloat(minimo) || 1,
                 id
             );
             req.flash('success', 'Producto actualizado correctamente');
